Refaktoroi organisaationSähköpostit henkilo->userdirectory
